backup/requirements.txt
# AI Agent Dependencies
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
typing-extensions>=4.0.0

# Weather and Email Dependencies
httpx>=0.24.0

# GCP Dependencies
google-cloud-scheduler>=2.0.0
google-cloud-functions>=1.0.0
google-cloud-logging>=3.0.0

# Additional dependencies
python-dateutil>=2.8.0
flask>=2.0.0 